+++ src/phylox/constants.py
@@ -1,0 +1,2 @@
+LABEL_ATTR = "label"
+LENGTH_ATTR = "length"