+++ src/phylox/dinetwork.py
@@ -2,10 +2,11 @@ import random
 
 import networkx as nx
 
-LABEL_ATTR = "label"
+from phylox.cherrypicking.base import CherryPickingMixin
+from phylox.constants import LABEL_ATTR, LENGTH_ATTR
 
 
-class DiNetwork(nx.DiGraph):
+class DiNetwork(nx.DiGraph, CherryPickingMixin):
     def __init__(self, *args, **kwargs):
         edges = kwargs.get("edges", [])
         super().__init__(edges, *args, **kwargs)
@@ -15,15 +16,36 @@ class DiNetwork(nx.DiGraph):
             self.nodes[label[0]][LABEL_ATTR] = label[1]
             self.label_to_node_dict[label[1]] = label[0]
 
+    def _clear_cached(self):
+        for attr in ["_leaves", "_reticulations", "_roots", "_reticulation_number"]:
+            if hasattr(self, attr):
+                delattr(self, attr)
+
     @classmethod
     def from_newick(cls, newick):
         pass
 
+    def _set_leaves(self):
+        self._leaves = set([node for node in self.nodes if self.is_leaf(node)])
+        return self._leaves
+
     @property
     def leaves(self):
         if not hasattr(self, "_leaves"):
-            self._leaves = set([node for node in self.nodes if self.is_leaf(node)])
+            self._set_leaves()
         return self._leaves
+
+    def _set_reticulations(self):
+        self._reticulations = set(
+            [node for node in self.nodes if self.is_reticulation(node)]
+        )
+        return self._reticulations
+
+    @property
+    def reticulations(self):
+        if not hasattr(self, "_retculations"):
+            self._set_reticulations()
+        return self._reticulations
 
     def _set_roots(self):
         self._roots = set([node for node in self.nodes if self.is_root(node)])