+++ tests/cherrypicking/test_tree_child_sequence.py
@@ -30,13 +30,12 @@ class TestTreeChildContainment(unittest.TestCase):
     def test_simple_cherry_picking_labelled(self):
         network1 = DiNetwork(
             edges=[(0, 1), (1, 2), (1, 3), (2, 3), (2, 4), (3, 5)],
-            labels=[[4,"a"], [5,"b"]]
+            labels=[[4, "a"], [5, "b"]],
         )
         network2 = DiNetwork(
-            edges=[(0, 1), (1, 2), (1, 3)],
-            labels=[[2,"a"], [3,"b"]]
+            edges=[(0, 1), (1, 2), (1, 3)], labels=[[2, "a"], [3, "b"]]
         )
         result = tree_child_network_contains(network1, network2)
         self.assertFalse(result)
         result = tree_child_network_contains(network1, network2, labels=True)
-        self.assertTrue(result)        
+        self.assertTrue(result)