+++ tests/test_dinetwork.py
@@ -1,6 +1,7 @@
 import unittest
 
-from phylox import LABEL_ATTR, DiNetwork
+from phylox import DiNetwork
+from phylox.constants import LABEL_ATTR
 
 
 class TestDiNetwork(unittest.TestCase):